moving helper around, putting conditional logic around display
