+++ src/components/ItemDetail/LineEntryHelper.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React from "react";
+import { Link } from "react-router-dom";
 
 export function calculatedLink(label, i, urls = null) {
   if (urls) {
